msleep() with hrtimers
The OLPC folks and I recently discovered something interesting: on a
HZ=100 system, a call to msleep(1) will delay for about 20ms.  The
combination of jiffies timekeeping and rounding up means that the
minimum delay from msleep will be two jiffies, never less.  That led to
multi-second delays in a driver which does a bunch of short msleep()
calls and, in response, a change to mdelay(), which will come back in
something closer to the requested time.

Here's another approach: a reimplementation of msleep() and
msleep_interruptible() using hrtimers.  On a system without real
hrtimers this code will at least drop down to single-jiffy delays much
of the time (though not deterministically so).  On my x86_64 system with
Thomas's hrtimer/dyntick patch applied, msleep(1) gives almost exactly
what was asked for.

jon

Use hrtimers for msleep() and msleep_interruptible()

Signed-off-by: Jonathan Corbet <corbet@lwn.net>